Fixed player animations disabled secondary box colliders on all lasers to prevent enemies from flying off the level added more to level 3 fixed bugs when player is cloned

+++ Assets/Standard Assets/Scripts/GameManager.cs
@@ -11,7 +11,9 @@ public static class GameManager {
 	public static void RestartLevel(){
 		//remove only temporary items
 		RemoveItemsFromPlayer(ItemDatabase.GetTemporaryList());
-		GameObject.Find("Player1").GetComponent<PlayerController>().ResetHealth();
+		GameObject player = GetPlayerObject();
+		if (player!= null)
+			player.GetComponent<PlayerController>().ResetHealth();
 		Application.LoadLevel(Application.loadedLevel);
 	}
 	public static void PauseGameToggle(){
@@ -19,14 +21,19 @@ public static class GameManager {
 	}
 	public static void AddHealthToPlayer(int value){
 		try{
-			GameObject.Find("Player1").GetComponent<PlayerController>().IncreaseHealth(value);
+			GameObject player = GetPlayerObject();
+			if (player!= null)
+				player.GetComponent<PlayerController>().IncreaseHealth(value);
 		}catch (Exception e){
 			Debug.Log(e.Message);
 		}
 	}
 	public static void AddPowerUpToPlayer(Item.PowerUpType powerUp){
 		try{
-			GameObject.Find("Player1").GetComponent<PlayerController>().AddPowerUp(powerUp);
+			GameObject player = GetPlayerObject();
+			if (player!= null)
+				player.GetComponent<PlayerController>().AddPowerUp(powerUp);
+
 		}catch (Exception e){
 			Debug.Log(e.Message);
 		}
@@ -56,13 +63,23 @@ public static class GameManager {
 	}
 	private static void RemoveItemsFromPlayer(List<Item> list){
 		List<Item> removeFromPlayer = new List<Item>(list);
-		if (removeFromPlayer.Count > 0)
-		foreach(Item item in removeFromPlayer){
-			if (item.ItemPowerUpType != Item.PowerUpType.None)
-				GameObject.Find("Player1").GetComponent<PlayerController>().RemovePowerUp(item.ItemPowerUpType);
-			
-			ItemDatabase.RemoveItem(item);
+		if (removeFromPlayer.Count > 0){
+			GameObject player = GetPlayerObject();
+			if (player!= null)
+				foreach(Item item in removeFromPlayer){
+					if (item.ItemPowerUpType != Item.PowerUpType.None)
+						player.GetComponent<PlayerController>().RemovePowerUp(item.ItemPowerUpType);
+					
+					ItemDatabase.RemoveItem(item);
+				}
 		}
+	}
+
+	private static GameObject GetPlayerObject(){
+		GameObject player = GameObject.Find("Player1");
+		if (player == null)
+			player = GameObject.Find("Player1" + GameResources.ObjectWasCloned);
+		return player;
 	}
 	
 }
